service: Add a systemd unit for D-Bus activation
On systems using systemd --user and
dbus-daemon --session --systemd-activation, this means that dconf
is correctly placed in its own cgroup, instead of being part of
dbus.service's cgroup. This allows it to be controlled by
systemctl --user, have resource limits set and so on.

On other systems (in particular those where systemd is not used at all),
the SystemdService field is ignored and dbus-daemon will continue to
start dconf-service as its own child process ("traditional activation"),
making this change harmless to apply.

Resolves: https://gitlab.gnome.org/GNOME/dconf/-/issues/24
Signed-off-by: Simon McVittie <smcv@collabora.com>